+++ packages/dock-blockchain-modules/src/accumulator/module.js
@@ -12,7 +12,7 @@ import {
   DockAccumulatorIdIdent,
 } from '@docknetwork/credential-sdk/types';
 import { option, withProp } from '@docknetwork/credential-sdk/types/generic';
-import { AbstractAccumulatorModule } from '@docknetwork/credential-sdk/modules/abstract/accumulator';
+import { AbstractAccumulatorModule } from '@docknetwork/credential-sdk/modules/abstract';
 import DockInternalAccumulatorModule from './internal';
 import { withParamsAndPublicKeys } from '../common';