+++ packages/dock-blockchain-modules/src/anchor/index.js
@@ -1,0 +1,2 @@
+export { default as DockAnchorModuleInternal } from './internal';
+export { default as DockAnchorModule } from './module';