+++ packages/dock-blockchain-modules/src/did/index.js
@@ -1,0 +1,3 @@
+export * from './actions';
+export { default as DockDIDModuleInternal } from './internal';
+export { default as DockDIDModule } from './module';