+++ packages/dock-blockchain-modules/src/offchain-signatures/bbs.js
@@ -5,7 +5,7 @@ import DockOffchainSignaturesModule from './module';
 import DockInternalOffchainSignaturesModule from './internal';
 
 /** Class to write `BBS` parameters and keys on chain */
-export default class BBSModule extends DockOffchainSignaturesModule {
+export default class DockBBSModule extends DockOffchainSignaturesModule {
   static DockOnly = class DockInternalBBSModuleOverrides extends DockInternalOffchainSignaturesModule {
     static PublicKey = BBSPublicKey;