+++ packages/dock-blockchain-modules/src/offchain-signatures/index.js
@@ -1,0 +1,5 @@
+export * from './actions';
+export { default as DockBBSModule } from './bbs';
+export { default as DockBBSPlusModule } from './bbs-plus';
+export { default as DockPSModule } from './bbs-plus';
+export { default as DockOffchainSignaturesModule } from './module';